pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "objlib"
version = "0.1.0"
description = "Objectivism Library semantic search scanner and metadata extractor"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "google-genai>=1.63.0",
    "tenacity>=9.1",
    "aiosqlite>=0.22",
    "keyring>=25.0",
    "mistralai>=1.0",
    "pydantic>=2.0",
    "aiolimiter>=1.1",
    "rapidfuzz==3.6.1",
    "textual>=5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "pyfakefs>=5.7",
]

[project.scripts]
objlib = "objlib.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/objlib"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["src/objlib/cli.py", "src/objlib/__pycache__/*"]

[tool.coverage.report]
fail_under = 80
